refactor(networks): type networks config and unify NetworkOption
