+++ packages/frontend/src/constants/networks.ts
@@ -1,9 +1,18 @@
 export const NETWORKS_ENUM = {
   MAINNET: 'mainnet',
   TESTNET: 'testnet'
+} as const
+
+export interface NetworkOption {
+  name: string
+  subname: string
+  disabled: boolean
+  explorerBaseUrl: string | undefined
+  l1explorerBaseUrl: string | undefined
+  dataContractPE: string | undefined
 }
 
-export const NETWORK_OPTIONS = {
+export const NETWORK_OPTIONS: Record<string, NetworkOption> = {
   [NETWORKS_ENUM.MAINNET]: {
     name: NETWORKS_ENUM.MAINNET,
     subname: '',
